Feat: User Delete test fail by user not found

+++ java-backend-code-challenge/src/main/java/com/codechallenge/application/usecase/UserDelete.java
@@ -2,6 +2,7 @@ package com.codechallenge.application.usecase;
 
 import com.codechallenge.application.domain.User;
 import com.codechallenge.application.ports.driven.UserRepository;
+import com.codechallenge.application.usecase.exception.UserNotFoundException;
 
 import java.util.Optional;
 
@@ -14,7 +15,9 @@ public class UserDelete {
 
     public void execute(String username){
         Optional<User> searchedUser = repository.getById(username);
-
+        if (searchedUser.isEmpty()) {
+            throw new UserNotFoundException("There isn't an user with that ID");
+        }
         repository.deleteUser(searchedUser.get());
     }
 }
